--- Maestri.Back/Common/Contracts/WebHooks/CreatePixSlipWebHook.py
@@ -1,9 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass(frozen = True, order = True)
-class CreatePixSlipWebHook():
-    hook_url: str
-
-    @staticmethod
-    def create(data: dict):
-        return CreatePixSlipWebHook(data["hook_url"])
