--- Maestri.Back/Common/Enums/RequestType.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-class RequestType (Enum):
-    GET = 'GET'
-    POST = 'POST'
-    PUT = 'PUT'
-    DELETE = 'DELETE'
-    PATCH = 'PATCH'
