--- Maestri.Back/Model/Client/ClientTokenClass.py
@@ -1,5 +1,0 @@
-
-class ClientTokenClass:
-    def __init__(self):
-        self.id = "-id-"
-        self.secret = "-secret-"
