+++ Maestri.Back/common/contracts/hook/create_pix_slip_hook.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True, order=True)
+class CreatePixSlipWebHook:
+    hook_url: str
+
+    @staticmethod
+    def create(data: dict):
+        return CreatePixSlipWebHook(data["hook_url"])
