+++ Maestri.Back/common/enums/person_enums.py
@@ -1,0 +1,40 @@
+from enum import Enum
+
+
+class PersonType(Enum):
+    PF = "FISICA"
+    PJ = "JURIDICA"
+
+    @staticmethod
+    def get_field_dict():
+        return {PersonType.PF: "cpf", PersonType.PJ: "cnpj"}
+
+
+class UF(Enum):
+    AC = "AC"
+    AL = "AL"
+    AP = "AP"
+    AM = "AM"
+    BA = "BA"
+    CE = "CE"
+    DF = "DF"
+    ES = "ES"
+    GO = "GO"
+    MA = "MA"
+    MT = "MT"
+    MS = "MS"
+    MG = "MG"
+    PA = "PA"
+    PB = "PB"
+    PR = "PR"
+    PE = "PE"
+    PI = "PI"
+    RJ = "RJ"
+    RN = "RN"
+    RS = "RS"
+    RO = "RO"
+    RR = "RR"
+    SC = "SC"
+    SP = "SP"
+    SE = "SE"
+    TO = "TO"
