+++ Maestri.Back/common/enums/request_type.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class RequestType(Enum):
+    GET = "GET"
+    POST = "POST"
+    PUT = "PUT"
+    DELETE = "DELETE"
+    PATCH = "PATCH"
